requirements.txt
streamlit==1.24.1
streamlit_chat
openai==0.27.8
streamlit_extras
psycopg2-binary
shortuuid
watchdog